+++ CHANGELOG.md
@@ -38,8 +38,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 ## Unreleased
 
 ### Improvements
-* 
+
 * [#297](https://github.com/babylonlabs-io/vigilante/pull/297) chore: unbonding watcher block metrics
+* [#298](https://github.com/babylonlabs-io/vigilante/pull/298) chore: log binary version
+* [#294](https://github.com/babylonlabs-io/vigilante/pull/294) chore: version cmd
+* [#296](https://github.com/babylonlabs-io/vigilante/pull/296) chore: tm retry container start
 
 ## v0.23.0